phy: exynos5-usbdrd: use GENMASK and FIELD_PREP for Exynos5 PHY registers
Exynos850 and Exynos9 (GS101) DRD PHY registers use GENMASK for masks
and FIELD_PREP for writing values to registers.

Change the Exynos5 registers to follow this approach as well as much as
possible. This patch doesn't introduce any fixes or functional changes,
it's merely an attempt to introduce some uniformity and consistency in the
driver.

The CRPORT SuperSpeed control registers have been exempted from the
change.

Signed-off-by: methanal <baclofen@tuta.io>